+++ packages/create-svelte/templates/default/src/routes/sverdle/+page.svelte
@@ -78,6 +78,11 @@
 
 <svelte:window on:keydown={keydown} />
 
+<svelte:head>
+	<title>Sverdle</title>
+	<meta name="description" content="A Wordle clone written in SvelteKit" />
+</svelte:head>
+
 <form
 	method="POST"
 	action="?/enter"
@@ -210,7 +215,7 @@
 		height: 100%;
 		display: flex;
 		flex-direction: column;
-		justify-content: start;
+		justify-content: flex-start;
 	}
 
 	.grid .row {